+++ hata/ext/slash/slasher.py
@@ -1276,6 +1276,8 @@ class Slasher(EventHandlerBase):
             Whether the command should be deleted from Discord when removed.
         allow_by_default : `None`, `bool`, `tuple` of (`None`, `bool`, `Ellipsis`), Optional
             Whether the command is enabled by default for everyone who has `use_application_commands` permission.
+        allow_in_dm : `None`, `bool`, `tuple` of (`None`, `bool`, `Ellipsis`), Optional
+            Whether the command can be used in private channels (dm).
         custom_id : `str`, (`list`, `set`) of `str`, `tuple` of (`str`, (`list`, `set`) of `str`)
             Custom id to match by the component command.
         allowed_mentions : `None`, `str`, ``UserBase``, ``Role``, ``AllowedMentionProxy``, \